requirements.txt
Pillow
PyQT6
requests
fitz
PyMuPDF